construct head

+++ include/log/severity.hpp
@@ -2,6 +2,6 @@
 
 namespace miu::log {
 
-enum struct severity : uint32_t { DEBUG, INFO, WARN, ERROR, MAX };
+enum struct severity : uint8_t { DEBUG, INFO, WARN, ERROR, MAX };
 
 }
